refactored tasks into their own folder
