+++ README.md
@@ -75,11 +75,12 @@ In this example, we're using three `Image Description` nodes to describe the giv
 
 ## Recent Changelog
 
-### July 15, 2024
+### July 16, 2024
 * Added `max_subtasks` parameter to Create Agent node and Griptape Run: Toolkit Task node to allow users to decide how many subtasks they want the agent to try. Default is 20.
 * Added `max_attemnpts_on_fail` parameter to all Config nodes to allow the user to determine the number of retries they want when an agent fails. This maps to the `max_attempts` parameter in the Griptape Framework.
+* **NewNode**: Added AzureOpenAI Config node. To use this, you'll need to set up your Azure endpoint and get API keys. The two environment variables required are `AZURE_OPENAI_ENDPOINT` and `AZURE_OPENAI_API_KEY`. You will also require a [deployment name](https://learn.microsoft.com/en-us/azure/ai-services/openai/how-to/switching-endpoints). This is available in [Azure OpenAI Studio](https://oai.azure.com/)
 * Updated README
-* 
+
 ### July 12, 2024
 * Updated to Griptape v0.28.2
 * **New Node** Griptape Config: OpenAI Compatable node. Allows you to connect to services like https://www.ohmygpt.com/ which are compatable with OpenAi's api.
@@ -146,6 +147,8 @@ ANTHROPIC_API_KEY=
 VOYAGE_API_KEY=
 GRIPTAPE_API_KEY=
 HUGGINGFACE_HUB_ACCESS_TOKEN=
+AZURE_OPENAI_ENDPOINT=
+AZURE_OPENAI_API_KEY=
 ```
 
 You can get the appropriate API keys from these respective sites:
@@ -164,6 +167,7 @@ You can get the appropriate API keys from these respective sites:
 * ANTHROPIC_API_KEY: https://console.anthropic.com/settings/keys
 * VOYAGE_API_KEY: https://dash.voyageai.com/
 * HUGGINGFACE_HUB_ACCESS_TOKEN: https://huggingface.co/settings/tokens
+* AZURE_OPENAI_ENDPOINT & AZURE_OPENAI_API_KEY: https://learn.microsoft.com/en-us/azure/ai-services/openai/how-to/switching-endpoints
 
 #### 4. Install Griptape-ComfyUI
