+++ nodes/__init__.py
@@ -1,5 +1,5 @@
-import sys
 import os
+import sys
 
 # Get the directory of the current script (__file__ is the path to the current file)
 current_dir = os.path.dirname(__file__)
