+++ nodes/tasks/audio_transcription_task.py
@@ -1,0 +1,89 @@
+from textwrap import dedent
+
+from griptape.drivers import (
+    DummyAudioTranscriptionDriver,
+    OpenAiAudioTranscriptionDriver,
+)
+from griptape.engines import (
+    AudioTranscriptionEngine,
+)
+from griptape.loaders import AudioLoader
+from griptape.structures import Pipeline
+from griptape.tasks import (
+    AudioTranscriptionTask,
+)
+from griptape.utils import load_file
+
+from ...py.griptape_config import get_config
+from ..base_audio_task import gtUIBaseAudioTask
+
+default_prompt = "{{ input_string }}"
+OPENAI_API_KEY = get_config("env.OPENAI_API_KEY")
+
+
+class gtUIAudioTranscriptionTask(gtUIBaseAudioTask):
+    DESCRIPTION = "Transcribe an audio file."
+    CATEGORY = "Griptape/Audio"
+
+    RETURN_TYPES = ("STRING",)
+    RETURN_NAMES = ("OUTPUT",)
+
+    def run(self, **kwargs):
+        audio = kwargs.get("audio", None)
+        audio_filepath = kwargs.get("audio_filepath", None)
+        driver = kwargs.get("driver", None)
+
+        audio_artifact = None
+        if audio:
+            audio = self.save_audio_tempfile(audio)[0]
+        elif audio_filepath:
+            audio = audio_filepath
+        else:
+            return ("There is no audio file.",)
+
+        try:
+            audio_artifact = AudioLoader().load(load_file(audio))
+        except Exception as e:
+            print(f"Error loading audio file: {e}")
+        if audio_artifact:
+            if not driver:
+                audio_transcription_driver = OpenAiAudioTranscriptionDriver(
+                    model="whisper-1"
+                )
+            else:
+                audio_transcription_driver = driver
+
+            # If the driver is a DummyAudioTranscriptionDriver we'll return a nice error message
+            if isinstance(audio_transcription_driver, DummyAudioTranscriptionDriver):
+                return (
+                    dedent(
+                        """
+                    I'm sorry, this agent doesn't have access to a valid AudioTranscriptionDriver.
+                    You might want to try using a different Agent Configuration.
+
+                    Reach out for help on Discord (https://discord.gg/gnWRz88eym) if you would like some help.
+                    """,
+                    ),
+                )
+            engine = AudioTranscriptionEngine(
+                audio_transcription_driver=audio_transcription_driver
+            )
+            # prompt_text = self.get_prompt_text(STRING, input_string)
+
+            task = AudioTranscriptionTask(
+                input=lambda _: audio_artifact,
+                audio_transcription_engine=engine,
+            )
+
+            # if deferred_evaluation:
+            #     return ("Audio Transcription Task created", task)
+            pipeline = Pipeline()
+            pipeline.add_task(task)
+            try:
+                result = pipeline.run()
+            except Exception as e:
+                print(e)
+            output = result.output_task.output.value
+        else:
+            output = "No audio provided"
+        return (output,)
