+++ nodes/tasks/text_summary_task.py
@@ -1,0 +1,32 @@
+from griptape.tasks import (
+    TextSummaryTask,
+)
+
+from ...py.griptape_config import get_config
+from ..agent.agent import gtComfyAgent as Agent
+from .base_task import gtUIBaseTask
+
+default_prompt = "{{ input_string }}"
+OPENAI_API_KEY = get_config("env.OPENAI_API_KEY")
+
+
+class gtUITextSummaryTask(gtUIBaseTask):
+    DESCRIPTION = "Summarize a text prompt."
+
+    def run(self, **kwargs):
+        STRING = kwargs.get("STRING")
+        input_string = kwargs.get("input_string", None)
+        agent = kwargs.get("agent", None)
+
+        if not agent:
+            agent = Agent()
+        prompt_text = self.get_prompt_text(STRING, input_string)
+        task = TextSummaryTask(prompt_text)
+        # if deferred_evaluation:
+        #     return ("Text Summary Task Created", agent, task)
+        try:
+            agent.add_task(TextSummaryTask(prompt_text))
+        except Exception as e:
+            print(e)
+        result = agent.run()
+        return (result.output_task.output.value, agent, task)
