+++ nodes/tasks/toolkit_task.py
@@ -1,0 +1,64 @@
+from griptape.tasks import (
+    ToolkitTask,
+)
+
+from ...py.griptape_config import get_config
+from ..agent.agent import gtComfyAgent as Agent
+from .base_task import gtUIBaseTask
+
+default_prompt = "{{ input_string }}"
+OPENAI_API_KEY = get_config("env.OPENAI_API_KEY")
+
+
+class gtUIToolkitTask(gtUIBaseTask):
+    DESCRIPTION = "Provide a list of tools, and have the agent decide which of them to use utilizing Chain of Thought."
+
+    @classmethod
+    def INPUT_TYPES(cls):
+        inputs = super().INPUT_TYPES()
+
+        # Update optional inputs to include 'tool' and adjust others as necessary
+        inputs["required"].update(
+            {
+                "max_subtasks": ("INT", {"default": 20}),
+            }
+        )
+        inputs["optional"].update(
+            {
+                "tools": ("TOOL_LIST",),
+            }
+        )
+        return inputs
+
+    def run(self, **kwargs):
+        STRING = kwargs.get("STRING")
+        tools = kwargs.get("tools", [])
+        input_string = kwargs.get("input_string", None)
+        agent = kwargs.get("agent", None)
+        max_subtasks = kwargs.get("max_subtasks", 5)
+        prompt_text = self.get_prompt_text(STRING, input_string)
+
+        if len(tools) == 0:
+            return super().run(STRING, input_string, agent)
+
+        if prompt_text.strip() == "":
+            return ("No prompt provided", agent)
+        # if the tool is provided, keep going
+        if not agent:
+            agent = Agent()
+
+        model, simple_model = agent.model_check()
+        if simple_model:
+            response = agent.model_response(model)
+            return (response, agent)
+
+        task = ToolkitTask(prompt_text, tools=tools, max_subtasks=max_subtasks)
+        # if deferred_evaluation:
+        #     return ("Toolkit Task Created.", task)
+        try:
+            agent.add_task(task)
+        except Exception as e:
+            print(e)
+
+        result = agent.run()
+        return (result.output_task.output.value, agent)
